Restructure project Add Eclipse settings if you prefer to use an IDE and configure it to use the default Eclipse formatter. Also fixed formatting in the Rule class and moved the subclasses into the parent because you can only have 1 top level class per Java source file by convention.
