+++ .settings/org.eclipse.jdt.ui.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+formatter_profile=org.eclipse.jdt.ui.default.eclipse_profile
+formatter_settings_version=12
