+++ src/edu/psu/cmpsc/teamone/Instructor.java
@@ -1,0 +1,5 @@
+package edu.psu.cmpsc.teamone;
+
+public class Instructor {
+
+}
