+++ src/edu/psu/cmpsc/teamone/Meeting.java
@@ -1,0 +1,5 @@
+package edu.psu.cmpsc.teamone;
+
+public class Meeting {
+
+}
