+++ src/edu/psu/cmpsc/teamone/Schedule.java
@@ -1,0 +1,5 @@
+package edu.psu.cmpsc.teamone;
+
+public class Schedule {
+
+}
