+++ src/edu/psu/cmpsc/teamone/Section.java
@@ -1,0 +1,5 @@
+package edu.psu.cmpsc.teamone;
+
+public class Section {
+
+}
